+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- add `ldap.py` for an ldap endpoint for getting a person by string
+- added parameterized tests for it in `test_ldap.py`
 
 ### Changes